mamunnnn

+++ Source/KGERP.Data/Models/Accounting_CostCenterType.cs
@@ -14,7 +14,7 @@ namespace KGERP.Data.Models
     
     public partial class Accounting_CostCenterType
     {
-        public int CostCenterTypeId { get; set; }
+        public long CostCenterTypeId { get; set; }
         public string Name { get; set; }
         public int CompanyId { get; set; }
         public string CreatedBy { get; set; }
